Accept sampling data with variable size fields
We used hardcoded sizes of FLOW_SAMPLER_ID and
FLOW_SAMPLER_RANDOM_INTERVAL fields when retrieving sampling data.
Some routers use non-standard sizes for those fields though.

+++ src/types.h
@@ -122,7 +122,7 @@ struct stream_id
 struct sampler_id
 {
     device_id did;
-    uint16_t sid;
+    uint32_t sid;
 };
 
 template <>
